Simplify JsonCodec/JsonCodecable and integrate dependency injection
We now use a codec class to serialize/deserialize objects of a certain
type, and the codec class is provided with a service container to
instantiate any resources it needs.  This persistent codec class can
ensure that (for example) singleton objects are resolved correctly
in the global state, without explicitly exposing those mechanisms
to the codec.

A helper JsonCodecableTrait allows *most* users to simply define
the toJsonArray() and newFromJsonArray() methods as usual; only
advanced users need to explicitly define a jsonClassCodec() method.