Rewrite class hint suffix system to use Hint objects
Using a strongly-typed Hint object allows us to retain the strong
checking of class-string values, and provides a more extensible system
that allows additional hint modification types.  This commit adds
Hint::STDCLASS (like Hint::ARRAY and the `[]` suffix, but cast to
object), and Hint::INHERITED (for superclass codecs that can
instantiate their subclasses).

This is a breaking change because the method parameter signatures have
been broadened.

Change-Id: Ie803f1cadcc1e4fc0f789f8e4d4b978a9486b765